requirements.txt
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
urllib3==2.5.0
